[SPOT-293][REFACTOR] 알림 기능 유즈케이스 별 컨트롤러, 서비스 코드 분리
