+++ src/main/java/com/example/spot/notification/application/impl/GetNotificationUseCaseImpl.java
@@ -1,32 +1,32 @@
-package com.example.spot.notification.application.legacy;
+package com.example.spot.notification.application.impl;
 
 import com.example.spot.common.api.code.status.ErrorStatus;
 import com.example.spot.common.api.exception.GeneralException;
+import com.example.spot.notification.application.GetNotificationUseCase;
 import com.example.spot.notification.domain.Notification;
+import com.example.spot.notification.domain.NotificationRepository;
 import com.example.spot.notification.domain.enums.NotifyType;
-import com.example.spot.notification.presentation.legacy.dto.notification.NotificationResponseDTO.NotificationListDTO;
-import com.example.spot.notification.presentation.legacy.dto.notification.NotificationResponseDTO.NotificationListDTO.NotificationDTO;
-import com.example.spot.notification.presentation.legacy.dto.notification.NotificationResponseDTO.StudyNotificationListDTO;
-import com.example.spot.notification.presentation.legacy.dto.notification.NotificationResponseDTO.StudyNotificationListDTO.StudyNotificationDTO;
+import com.example.spot.notification.presentation.dto.response.NotificationResponseDTO.NotificationListDTO;
+import com.example.spot.notification.presentation.dto.response.NotificationResponseDTO.NotificationListDTO.NotificationDTO;
+import com.example.spot.notification.presentation.dto.response.NotificationResponseDTO.StudyNotificationListDTO;
+import com.example.spot.notification.presentation.dto.response.NotificationResponseDTO.StudyNotificationListDTO.StudyNotificationDTO;
 import java.util.ArrayList;
 import java.util.List;
-
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
-
-import com.example.spot.notification.domain.NotificationRepository;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
-public class NotificationQueryServiceImpl implements NotificationQueryService {
+public class GetNotificationUseCaseImpl implements GetNotificationUseCase {
 
     private final NotificationRepository notificationRepository;
 
     /**
      * 회원이 참가 신청한 스터디에 대한 알림을 조회합니다.
+     *
      * @param memberId 사용자 ID
      * @param pageable 페이징 정보
      * @return 참가 신청한 스터디에 대한 알림 목록
@@ -36,11 +36,12 @@ public class NotificationQueryServiceImpl implements NotificationQueryService {
     public StudyNotificationListDTO getAllAppliedStudyNotification(Long memberId, Pageable pageable) {
         // 특정 회원이 참가 신청한 스터디에 대한 알림 조회
         List<Notification> notifications = notificationRepository.findByMemberIdAndTypeAndIsChecked(
-            memberId, pageable, NotifyType.STUDY_APPLY, false);
+                memberId, pageable, NotifyType.STUDY_APPLY, false);
 
         // 알림이 존재하지 않는 경우
-        if (notifications.isEmpty())
+        if (notifications.isEmpty()) {
             throw new GeneralException(ErrorStatus._NOTIFICATION_NOT_FOUND);
+        }
 
         // DTO를 담을 리스트 생성
         List<StudyNotificationDTO> notificationDTOs = new ArrayList<>();
@@ -48,27 +49,29 @@ public class NotificationQueryServiceImpl implements NotificationQueryService {
         // 알림을 DTO로 변환하여 리스트에 추가
         notifications.forEach(notification -> {
             StudyNotificationDTO notificationDTO = StudyNotificationDTO.builder()
-                .notificationId(notification.getId())
-                .studyId(notification.getStudy().getId())
-                .createdAt(notification.getCreatedAt())
-                .type(notification.getType())
-                .studyTitle(notification.getStudy().getTitle())
-                .studyProfileImage(notification.getStudy().getProfileImage())
-                .isChecked(notification.getIsChecked())
-                .build();
+                    .notificationId(notification.getId())
+                    .studyId(notification.getStudy().getId())
+                    .createdAt(notification.getCreatedAt())
+                    .type(notification.getType())
+                    .studyTitle(notification.getStudy().getTitle())
+                    .studyProfileImage(notification.getStudy().getProfileImage())
+                    .isChecked(notification.getIsChecked())
+                    .build();
             notificationDTOs.add(notificationDTO);
         });
 
         // DTO 리스트를 반환
         return StudyNotificationListDTO.builder()
-            .notifications(notificationDTOs)
-            .totalNotificationCount((long) notificationDTOs.size())
-            .uncheckedNotificationCount((long) (int) notificationDTOs.stream().filter(notificationDTO -> !notificationDTO.getIsChecked()).count())
-            .build();
+                .notifications(notificationDTOs)
+                .totalNotificationCount((long) notificationDTOs.size())
+                .uncheckedNotificationCount((long) (int) notificationDTOs.stream()
+                        .filter(notificationDTO -> !notificationDTO.getIsChecked()).count())
+                .build();
     }
 
     /**
      * 회원에게 생성된 알림을 전체 조회합니다.
+     *
      * @param memberId 사용자 ID
      * @param pageable 페이징 정보
      * @return 알림 목록
@@ -79,11 +82,12 @@ public class NotificationQueryServiceImpl implements NotificationQueryService {
 
         // 특정 회원에게 생성된 알림을 조회
         List<Notification> notifications = notificationRepository.findByMemberIdAndTypeNot(
-            memberId, pageable, NotifyType.STUDY_APPLY);
+                memberId, pageable, NotifyType.STUDY_APPLY);
 
         // 알림이 존재하지 않는 경우
-        if (notifications.isEmpty())
+        if (notifications.isEmpty()) {
             throw new GeneralException(ErrorStatus._NOTIFICATION_NOT_FOUND);
+        }
 
         // DTO를 담을 리스트 생성
         List<NotificationDTO> notificationDTOs = new ArrayList<>();
@@ -106,12 +110,12 @@ public class NotificationQueryServiceImpl implements NotificationQueryService {
 
         // DTO 리스트를 반환
         return NotificationListDTO.builder()
-            .notifications(notificationDTOs)
-            .totalNotificationCount((long) notificationDTOs.size())
-            .uncheckedNotificationCount((long) (int) notificationDTOs.stream().filter(notificationDTO -> !notificationDTO.getIsChecked()).count())
-            .build();
+                .notifications(notificationDTOs)
+                .totalNotificationCount((long) notificationDTOs.size())
+                .uncheckedNotificationCount((long) (int) notificationDTOs.stream()
+                        .filter(notificationDTO -> !notificationDTO.getIsChecked()).count())
+                .build();
     }
-
 
 
 }
