+++ src/main/java/com/example/spot/notification/application/impl/ReadNotificationUseCaseImpl.java
@@ -1,0 +1,58 @@
+package com.example.spot.notification.application.impl;
+
+import com.example.spot.common.api.code.status.ErrorStatus;
+import com.example.spot.common.api.exception.GeneralException;
+import com.example.spot.notification.application.ReadNotificationUseCase;
+import com.example.spot.notification.domain.Notification;
+import com.example.spot.notification.domain.NotificationRepository;
+import com.example.spot.notification.presentation.dto.response.NotificationResponseDTO.NotificationProcessDTO;
+import java.util.Objects;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@Transactional
+@RequiredArgsConstructor
+public class ReadNotificationUseCaseImpl implements ReadNotificationUseCase {
+
+    private final NotificationRepository notificationRepository;
+
+    /**
+     * 알림을 읽음 처리 합니다. 이미 읽은 알림인 경우 예외를 발생시킵니다.
+     *
+     * @param memberId       사용자 ID
+     * @param notificationId 알림 ID
+     * @return 알림 읽음 처리 결과 및 처리 일시
+     * @throws GeneralException 알림이 존재하지 않거나 이미 읽음 처리된 경우
+     * @throws GeneralException 알림이 사용자에게 속하지 않는 경우
+     * @throws GeneralException 알림이 이미 읽음 처리된 경우
+     * @see NotificationProcessDTO
+     */
+    @Override
+    public NotificationProcessDTO readNotification(Long memberId, Long notificationId) {
+
+        // 알림 조회
+        Notification notification = notificationRepository.findById(notificationId).orElseThrow(
+                () -> new GeneralException(ErrorStatus._NOTIFICATION_NOT_FOUND));
+
+        // 알림이 사용자에게 속하지 않는 경우
+        if (!Objects.equals(notification.getMember().getId(), memberId)) {
+            throw new GeneralException(ErrorStatus._NOTIFICATION_IS_NOT_BELONG_TO_MEMBER);
+        }
+
+        // 이미 읽음 처리된 경우
+        if (notification.getIsChecked()) {
+            throw new GeneralException(ErrorStatus._NOTIFICATION_ALREADY_READ);
+        }
+
+        // 알림 읽음 처리
+        notification.markAsRead();
+
+        // 알림 읽음 처리 결과 반환
+        return NotificationProcessDTO.builder()
+                .isAccept(notification.getIsChecked())
+                .processedAt(notification.getUpdatedAt())
+                .build();
+    }
+}
