--- src/main/java/com/example/spot/notification/application/legacy/NotificationQueryService.java
@@ -1,15 +1,0 @@
-package com.example.spot.notification.application.legacy;
-
-import com.example.spot.notification.presentation.legacy.dto.notification.NotificationResponseDTO.NotificationListDTO;
-import com.example.spot.notification.presentation.legacy.dto.notification.NotificationResponseDTO.StudyNotificationListDTO;
-
-import org.springframework.data.domain.Pageable;
-
-public interface NotificationQueryService {
-
-    // 생성된 알림 전체 조회
-    NotificationListDTO getAllNotifications(Long memberId, Pageable pageable);
-
-    // 신청한 스터디 알림 전체 조회
-    StudyNotificationListDTO getAllAppliedStudyNotification(Long memberId, Pageable pageable);
-}
